Fix Zend_Form_ElementTest
The behaviour hasn't really changed because the intent of the test is to
use an array of arrays in setOptions. What has changed is that PHP 8.0
supports calling functions with named parameters and will throw a Fatal
if you try to access a parameter by name that does not exist.
Digits::__construct has no param named bar ergo we get a Fatal.

Signed-off-by: Alexander Wozniak <wozniak@oktopos.com>
Signed-off-by: Elan Ruusamäe <glen@pld-linux.org>