Removed trivial but problematic test
For whatever reason calling setFormatter on Zend_Log_Writer with an
object that does not fulfill the appropriate interface causes a fatal
error under PHP 5.4 that can't be caught.
These tests have been removed for the CI to completely run as the
only thing they really test is PHP's type checking and do not offer
any further insight.